+++ src/io/segment/android/models/EventProperties.java
@@ -1,7 +1,17 @@
 package io.segment.android.models;
 
+import org.json.JSONObject;
+
 public class EventProperties extends EasyJSONObject {
- 
+
+	public EventProperties() {
+		super();
+	}
+
+	public EventProperties(JSONObject json) {
+		super(json);
+	}
+	
 	public EventProperties(Object... kvs) {
 		super(kvs);
 	}