+++ src/io/segment/android/models/Traits.java
@@ -1,9 +1,16 @@
 package io.segment.android.models;
 
+import org.json.JSONObject;
+
 public class Traits extends EasyJSONObject {
 
+	public Traits() {}
+	
 	public Traits(Object... kvs) {
 		super(kvs);
 	}
 	
+	public Traits(JSONObject json) {
+		super(json);
+	}
 }