+++ tests/src/io/segment/android/providers/test/OmnitureProviderTest.java
@@ -3,20 +3,21 @@ package io.segment.android.providers.test;
 import io.segment.android.models.EasyJSONObject;
 import io.segment.android.provider.BaseProviderTest;
 import io.segment.android.provider.Provider;
-import io.segment.android.providers.MixpanelProvider;
+import io.segment.android.providers.OmnitureProvider;
 
-public class MixpanelProviderTest extends BaseProviderTest {
+public class OmnitureProviderTest extends BaseProviderTest {
 
 	@Override
 	public Provider getProvider() {
-		return new MixpanelProvider();
+		return new OmnitureProvider();
 	}
 
 	@Override
 	public EasyJSONObject getSettings() {
 		EasyJSONObject settings = new EasyJSONObject();
-		settings.put("token", "89f86c4aa2ce5b74cb47eb5ec95ad1f9");
+		settings.put("reportSuiteId", "xxxxxxxxx");
+		settings.put("trackingServer", "http://api.omniture.com");
 		return settings;
 	}
-	
+
 }